refactor: Move macOS collection view into a separate view
This is preparatory work for combining the macOS and iOS `SectionView` implementations.
